fix(CWL): use 23.224 Nightly new long source value

+++ CustomWhateverLoader/Helper/Extensions/CharaExt.cs
@@ -40,7 +40,7 @@ public static class CharaExt
                 added.vSourcePotential += added.GetSourcePotential(power);
             }
 
-            added.vSource += added.GetSourceValue(power, chara.LV, SourceValueType.Chara);
+            added.vSource += (int)added.GetSourceValue(power, chara.LV, SourceValueType.Chara);
 
             return added;
         }
